Rollup complete

+++ DesignDB_Library/Models/ReportSalesPriorityModel.cs
@@ -12,19 +12,19 @@ namespace DesignDB_Library.Models
         [ExcelExportPropertyAttribute]
         public string SalesPerson { get; set; }
         [ExcelExportPropertyAttribute]
-        public int P1Count { get; set; }
+        public decimal P1Count { get; set; }
         [ExcelExportPropertyAttribute]
         public decimal  P1Dollars { get; set; }
         [ExcelExportPropertyAttribute]
-        public int P2Count { get; set; }
+        public decimal P2Count { get; set; }
         [ExcelExportPropertyAttribute]
         public decimal P2Dollars { get; set; }
         [ExcelExportPropertyAttribute]
-        public int P3Count { get; set; }
+        public decimal P3Count { get; set; }
         [ExcelExportPropertyAttribute]
         public decimal P3Dollars { get; set; }
         [ExcelExportPropertyAttribute]
-        public int TotalCount { get; set; }
+        public decimal TotalCount { get; set; }
         [ExcelExportPropertyAttribute]
         public decimal TotalDollars { get; set; }
     }
